User seat overview continuing from sprint 1 (#83)
* Revert "Merge branch 'baseclass' into main"

This reverts commit bd7fba75a86e8d5d45ca62162923dd862e3c13cf, reversing
changes made to 0ede25b189a1d6b381c939e9344e6bda243f8f25.

* Added bar option

+ changed movieid to showid in the bar

* Added empty function

* Removed SetTakenSeats(showId) call

* Added empty seat logic

* seats user picks have to be connected and leave no 1 empty seats between them and other reservations

* Update TheaterBase.cs

* changed price as it was in main

* Changed row and col input checks as it was

* Fixing the left columns 1 empty seat issue

* adjusted left seat logic to begin when a seat appears

* changed prices

* Testing by filling the whole theater

Issues:

Groups can leave one seat empty at the left and then seat further on the right. Now there needs to be two empty seats. The only except is if there is no room on the right. Then everyone seats on the left.

* I refactored the code to make it more readable

* Update TheaterBase.cs

* Correctly fixed group reservations getting stuck

* Deleted comments

and removed unused method to check exactly two seats from the right side. Because it doesn't do anything

* Changed loop

Continue counting the seats until the selected seat is reached. Ensure the loop runs completely, as opposed to stopping prematurely as it did before. Make sure both the left and right sides are counted so that the selected seat is included when counting from both directions—left to right and right to left.